refactor: expose canonical language resolver
