+++ packages/sugar-high/lib/lang.js
