+++ packages/sugar-high/test/preset/c.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 import { tokenize } from '../..'
-import { c } from 'sugar-high/presets'
+import { c } from '../../lib/presets/index.js'
 import { getTokensAsString } from '../testing-utils'
 
 describe('tokenize - c preset', () => {
