+++ packages/sugar-high/test/preset/rust.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 import { tokenize } from '../..'
-import { rust } from 'sugar-high/presets'
+import { rust } from '../../lib/presets/index.js'
 import { getTokensAsString } from '../testing-utils'
 
 describe('tokenize - rust preset', () => {
